feat(speculative): make refusing a tree rollback the default too
Verifying a tree and keeping its accepted path are separate abilities, and
the branch so far only had a default for the first. `verify_forward_tree`
refuses unless a target opts in; `verify_finalize` had no counterpart, so a
target that opted into the forward would have inherited the linear rollback,
which trims a tail. Applied to a scattered accept set that keeps the wrong
entries rather than failing.

`verify_finalize_tree` takes the walk's path instead of an accept count and
refuses by default. For a linear tree the path is `[0, 1, ..., accepted]`,
the selection is a prefix, and the result has to be byte-identical to
`verify_finalize(accepted, block_size, captured)`. That equivalence is the
first thing to test once the round loop is wired, and a divergence there is
wiring rather than topology.

Accepting the forward and refusing the rollback is a real combination, not a
contradiction, and Gemma 4 is currently in it. Its sliding-window layers hold
`RotatingKVCache`, whose rollback rewinds `offset` and `idx` without moving
data, which a scattered selection cannot use, and which has no
`gather_positions`. So the two halves have to be probed as one capability
before a tree is drafted: once a tree forward has appended its block, a
refusal at the rollback leaves the cache in a state only a tree-aware
rollback can undo. `captured` is consumed rather than handed back for that
reason.

Qwen 3.5 overrides it to name its own cause rather than inherit the message,
matching what it already does for the forward. Unreachable while the forward
refuses first, and there so a forward that stops refusing cannot silently
inherit a rollback that cannot serve it.

Also fixes an `i32 -> i32` cast in tree_tests that fails `clippy -D warnings`.

Refs #1204

+++ src/lib/mlxcel-core/src/speculative/mtp/tree_tests.rs
@@ -37,7 +37,7 @@ fn causal_mask(n: usize) -> Vec<f32> {
 #[test]
 fn a_linear_tree_masks_exactly_as_a_causal_mask() {
     for k in 0..6 {
-        let drafts: Vec<i32> = (0..k).map(|i| 100 + i as i32).collect();
+        let drafts: Vec<i32> = (0..k).map(|i| 100 + i).collect();
         let tree = DraftTree::linear(7, &drafts);
         assert_eq!(
             tree.additive_mask(MASKED),
